Moved commands into their own object
Added read{byte,word,long} commands. Added a arg type checker: commands
can set a maximum size of argument and the input will be checked against
the list, which is setup with the checkcommand macro. The nocheckcommand
macro creates a command record with no check data, which means any
input (so long as it is still parser-able) will be allowed.