Added write{bytes,words,longs} commands
Extended the type checker to cope with "varargs": if the high bit is set in
the expected type, then it is the last type and a vararg; a list of the
same type until the end of the input.  This is used by the write commands
to output sequences of bytes, words, or longs, padding what's written to
fill the type set by the write command variant.

The monitor is now useable for excersising peripheral ICs! Have tested it
can talk to the CPLD ok already, next up a 6522 or something similar.

TODO: cleanup the type checker code (checktypes) a bit, it is a little
wasteful.