Some progress towards tasks
Can now start a single "task" by rte-ing into the code. Task code is
still running as supervisor though, which needs fixing.

Starting stack looks like this:

- Registers (unsure what to do about A7, but since the SP will come off
the task struct is probably redundant)
- Status register
- PC
- Format/Vector

The SP saved into the task struct is the top of this, so the "start a task"
code will first unstack the a/d registers and then rte.

The issue to resolve is usermode vs supervirsor. I'd like tasks to run
without supervisor priv, but that means the schedular will have to create
a supervisor struct (as above, bar registers) from data held in the task
structure, which will need to be copied out of the supervisor stack prior
to choosing the next task...