Switch the IO around a bit
a5 is now used to hold a "channel struct". The struct contains two sub
pointers for char get and put. Also made the byte/word/long to printable
characters write into the sub in a5, like strings and chars instead of
filling out an array. (This could still be done if needed if a sub was
written that wrote into an array.) So in essence instead of the dump
command printing into a ling long memory buffer it outputs the line
directly. This sould make it easier to form the data to output, though
of course buffers can still be filled and output with serputstr.