Rejig of debug functionality and redid hex parsing
Moved the debug macros into the monitor so the monitor can generate them
as well. This was needed when the hex parsing got redone, which was needed
because the debugger (gdb trap handling) will need to parse arbitary long
ASCII hex into a memory buffer for reading memory and the like.