Some progress on the disassembler
Understands ori.[wb] into %ccr and %sr, and most of the main ori opcode.
Next to work on is the register and PC with displacement indirect modes.
Will ignore everything later and including the 020 modes, for now.
Hopefully the base for the disassembler will allow me to make some fairly
rapid progress.