Move is done and seems to work nicely. :)
Added an alternative "front" to the main ea decoder that works for both
source and destination positions in the move opcode. Named the subs
moderightsub and modeleftsub, since this is what manual calls these
fields.

So Move is:

15:14 0b00
13:12 Size
11:9  Destination register
8:6   Destination mode
5:3   Source mode
2:0   Source register

d2 and d3 are used to hold the register and mode respectivly and are
set according to the type of extraction before the mode decoding happens.