Added jsr, jmp, mulu, muls, exg, and all the shifts
This should be all the 68000 instructions. exg is a bit of a pest as it
requires 3 instructions to be defined, one for each combination of addr
and data registers being exchanged. Shifts need 12 instructions defined.
There are 4 types of shift, and 3 modes: memory (always one shift left
or right), data reg where the amount is a 3 bit immediate, and data reg
where the amount is in another register. The shift direction is another
condition style instruction postfix.

Removed some usage of scaling the displacements when calcuating the
address of a string to print, since it is possible to simply shift
1 or 2 positions less, which will save some clocks and not require any
funky 020 instructions.

Next will be tidying up the code a bit.