Added PC relative displacement and index modes
This is a bit sneaky. Reuses the address reg modes, but calls a wrapper
around the addr reg printer which will look at d3 to check the mode for
111, the PC/absolute/immediate modes. If its 1111 it will print %pc,
otherwise it calls the normal address reg printer in a tail call.

This should be all of the 68000 ISA decoded. Still needs to be tarted up
to print the raw words between the address and the opcodes, and it needs
branch targets to be resolved to the PC+displacement+2.