Print the words from stream with decoded instructions
Like objdump, the instruction stream words are printed between the
address and the decoded instructions. This involves using a memeory
"io device" to hold the decoded instructions, which must be printed
AFTER the address and hex words, which are printed using the port passed
in via a5 at start. This memory io device is a little struct which
holds the instruction buffer and a current position count. There is
also a count for the characters needed to print the instruction words
as they are read in. This count is used to pad the left margin of the
decoded instructions to line them  up. After four words (eg a
move.b foo.l,bar.l (5 words needed), a newline is inserted and the
remaining words and the decoded instruction is printed on the next
line.

All told it seems to work quite well. Next up will be to finally move
the disassembler back to the flash.