Added the start of a system for working with breakpoints
Breakpoints (bps) use a struct holding:

1. Break point address
2. The word at that location

In this commit you can only list, create and remove breakpoints. The
word at the bp location will be replaced with a TRAP #n call when the
monitor is exited, and the old instruction stashed in the breakpoint's
struct. Then when the monitor is again entered (for any reason) all the
breakpoints will be examined and the instruction words will be replaced
with the "correct" word. Thus disassembly will show the "true"
instructions not the TRAPs, which are in place only when external code
is running.

In the next commit the disassembler and this new code will be moved
back to the flash, and the os-test code (resident in RAM) will become
the test code to run with breakpoints in place.